release profiles not quality profiles
