+++ pkg/bindata/templates/index.html
@@ -75,7 +75,7 @@
                                 <ul style="width:95%;" class="dropdown-menu bk-dark" aria-labelledby="triggers menu">
                                     <li><a class="nav-link text-grey" href="#triggers" onClick="swapNavigationTemplate('triggers')">- Open Triggers Page -</a></li>
                                     <li><a class="nav-link text-grey" onClick="triggerAction('cfsync')">Custom Formats</a></li>
-                                    <li><a class="nav-link text-grey" onClick="triggerAction('qpsync')">Quality Profiles</a></li>
+                                    <li><a class="nav-link text-grey" onClick="triggerAction('rpsync')">Release Profiles</a></li>
                                     <li><a class="nav-link text-grey" onClick="triggerAction('snapshot')">System Snapshot</a></li>
                                     <li><a class="nav-link text-grey" onClick="triggerAction('stuckitems')">Stuck Items</a></li>
                                     <li><a class="nav-link text-grey" onClick="triggerAction('dashboard')">Dashboard States</a></li>
