+++ pkg/bindata/templates/triggers.html
@@ -10,9 +10,9 @@
                                             </td>
                                         </tr>
                                         <tr>
-                                            <td>{{index .Expvar.TimerCounts "Starting Sonarr QP TRaSH sync."}}</td>
+                                            <td>{{index .Expvar.TimerCounts "Starting Sonarr RP TRaSH sync."}}</td>
                                             <td>
-                                                <a href="#triggers" onClick="triggerAction('qpsync')">Sync Quality Profiles</a></td><td>Syncs TRaSH quality profiles to Sonarr.
+                                                <a href="#triggers" onClick="triggerAction('rpsync')">Sync Release Profiles</a></td><td>Syncs TRaSH release profiles to Sonarr.
                                             </td>
                                         </tr>
                                         <tr>
